spark/work_baza/requirements.txt
cassandra-driver
bson